+++ src/components/TodoList.js
@@ -1,7 +1,8 @@
-import React from "react";
+import React, {useContext} from "react";
 import {Button, Table} from "reactstrap";
-
-function TodoList({todos, completeTodo, removeTodo}) {
+import {TodosContext} from "../contexts/TodosContext";
+function TodoList() {
+    const { todos, completeTodo, removeTodo} = useContext(TodosContext)
     return (
         <Table>
             <tbody>
